% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binary_vars}
\alias{binary_vars}
\title{Binary variables (as vector)}
\usage{
binary_vars(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of factor or character column names with two unique levels
}
\description{
Binary variables (as vector)
}
\section{Variables with binary (two levels)}{


This function is designed to quickly determine which variables have two
categorical levels. This is helpful for using \code{ggplot2::facet_wrap()}
or \code{ggplot2::facet_grid()}
}

\examples{
require(NHANES)
require(palmerpenguins)
binary_vars(df = NHANES::NHANES)
binary_vars(df = palmerpenguins::penguins)
# verify with str()
str(dplyr::select(NHANES::NHANES,
  dplyr::all_of(binary_vars(df = NHANES::NHANES))))
str(dplyr::select(palmerpenguins::penguins,
  dplyr::all_of(binary_vars(df = palmerpenguins::penguins))))
}
