% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{cat_app_inputs}
\alias{cat_app_inputs}
\title{Categorical app inputs}
\usage{
cat_app_inputs(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
character and factor column names
}
\description{
Categorical app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
cat_app_inputs(palmerpenguins::penguins)
cat_app_inputs(dplyr::starwars)
}
