% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deconstruct}
\alias{deconstruct}
\title{Deconstruct R objects}
\usage{
deconstruct(x, quotes = TRUE, console = TRUE)
}
\arguments{
\item{x}{R object passed to \code{dput()}}

\item{quotes}{include quotes in the output}

\item{console}{logical, used in the console? If \code{FALSE}, then output is printed
with \code{base::noquote()}. If \code{TRUE}, output is returned with \code{cat()}}
}
\value{
Deparsed object
}
\description{
Deconstruct R objects
}
\examples{
deconstruct(names(mtcars), quotes = FALSE)
deconstruct(names(mtcars), quotes = FALSE, console = FALSE)
deconstruct(x = names(mtcars), quotes = TRUE, console = TRUE)
}
