% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoInputsApp.R
\name{demoInputsApp}
\alias{demoInputsApp}
\title{Demo \verb{*app_inputs()} functions}
\usage{
demoInputsApp(data)
}
\arguments{
\item{data}{\code{app_data} can be any dataset (automatically loads with \code{NHANES::NHANES})}
}
\value{
demo shiny application
}
\description{
Demo \verb{*app_inputs()} functions
}
\examples{
require(unitTestAppPkg)
# demoInputsApp()
}
