% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_binary_checks}
\alias{df_binary_checks}
\title{Check for binary columns (helper)}
\usage{
df_binary_checks(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of binary columns
}
\description{
Check for binary columns (helper)
}
\examples{
require(palmerpenguins)
df_binary_checks(palmerpenguins::penguins, type = "fct")
}
