% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{facet_app_inputs}
\alias{facet_app_inputs}
\title{Facet app inputs}
\usage{
facet_app_inputs(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
character and factor column names with less than six levels
}
\description{
Facet app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
facet_app_inputs(palmerpenguins::penguins)
facet_app_inputs(dplyr::starwars)
}
