% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{facet_vars}
\alias{facet_vars}
\title{Facet variables (as vector)}
\usage{
facet_vars(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of factor or character column names with less than six
unique levels
}
\description{
Facet variables (as vector)
}
\section{Variables to use for facets}{


This function is designed to quickly determine which variables have an
appropriate number of categorical levels for using \code{ggplot2::facet_wrap()}
or \code{ggplot2::facet_grid()}
}

\examples{
require(dplyr)
require(NHANES)
facet_vars(df = dplyr::starwars)
facet_vars(df = NHANES::NHANES)

str(dplyr::select(dplyr::starwars,
  dplyr::all_of(facet_vars(df = dplyr::starwars))))
str(dplyr::select(NHANES::NHANES,
  dplyr::all_of(facet_vars(df = NHANES::NHANES))))
}
