% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_binary_vars}
\alias{fct_binary_vars}
\title{Factor binary variables}
\usage{
fct_binary_vars(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of factor column names with two unique levels
}
\description{
Factor binary variables
}
