% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{get_binary_checks_vec}
\alias{get_binary_checks_vec}
\title{Binary checks (vectors)}
\usage{
get_binary_checks_vec(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{type of column to return.}
}
\value{
TRUE/FALSE if binary
}
\description{
Binary checks (vectors)
}
\examples{
require(palmerpenguins)
get_binary_checks_vec(palmerpenguins::penguins$sex, type = "fct")
}
