% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{get_col_type_df}
\alias{get_col_type_df}
\title{Get column types (helper)}
\usage{
get_col_type_df(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return.
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"list"}: list
}}
}
\value{
columns of \code{type} (empty \code{tibble} is no columns of specified type
exist)
}
\description{
Return column in data by types
}
\examples{
require(dplyr)
get_col_type_df(dplyr::starwars, type = "chr")
get_col_type_df(dplyr::starwars, type = "list")
}
