% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_types.R
\name{get_col_type_tbl}
\alias{get_col_type_tbl}
\title{Get column types as tibble (helper)}
\usage{
get_col_type_tbl(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return.
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"list"}: list
}}
}
\value{
\code{tibble} of columns matching \code{type} (empty \code{tibble} is no columns
of specified type
exist)
}
\description{
Return column in data by types
}
\examples{
require(dplyr)
require(tidyr)
get_col_type_tbl(dplyr::starwars, type = "chr")
get_col_type_tbl(dplyr::starwars, type = "list")
get_col_type_tbl(tidyr::fish_encounters, "chr")
}
