% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_col_types.R
\name{get_col_types}
\alias{get_col_types}
\title{Get column types}
\usage{
get_col_types(df, type, return_tbl = TRUE)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"list"}: list
}}

\item{return_tbl}{logical, return tibble (\code{TRUE}) or named vector (\code{FALSE})}
}
\value{
named vector or \code{tibble} of columns matching \code{type} (empty vector or
\code{tibble} if no columns of specified type exist)
}
\description{
Return column in data by types
}
\examples{
require(dplyr)
require(tidyr)
get_col_types(dplyr::starwars, type = "chr")
get_col_types(dplyr::starwars, type = "chr", return_tbl = FALSE)
get_col_types(dplyr::starwars, type = "list")
get_col_types(dplyr::starwars, type = "list", return_tbl = FALSE)
get_col_types(tidyr::fish_encounters, "chr")
get_col_types(tidyr::fish_encounters, "chr", return_tbl = FALSE)
}
