% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_classes.R, R/get_col_types.R
\name{get_column_class}
\alias{get_column_class}
\title{Get column types}
\usage{
get_column_class(df, class, return_tbl = TRUE)

get_column_class(df, class, return_tbl = TRUE)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{class}{type of column to return
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"ord"}: factor (ordered)
\item \code{"list"}: list
}}

\item{return_tbl}{logical, return tibble (\code{TRUE}) or named vector (\code{FALSE})}

\item{type}{type of column to return
One of:
\itemize{
\item \code{"log"}: logical
\item \code{"dbl"}: double
\item \code{"int"}: integer
\item \code{"chr"}: character
\item \code{"fct"}: factor
\item \code{"ord"}: factor (ordered)
\item \code{"list"}: list
}}
}
\value{
named vector or \code{tibble} of columns matching the \code{type} (empty vector or
\code{tibble} if no columns of specified type exist)

named vector or \code{tibble} of columns matching the \code{class} (empty vector or
\code{tibble} if no columns of specified class exist)
}
\description{
Return column in data by types

Return column in data by class
}
\examples{
require(dplyr)
require(tidyr)
get_column_class(dplyr::starwars, class = "chr")
get_column_class(dplyr::starwars, class = "chr", return_tbl = FALSE)
get_column_class(dplyr::starwars, class = "list")
get_column_class(dplyr::starwars, class = "list", return_tbl = FALSE)
get_column_class(tidyr::fish_encounters, "chr")
get_column_class(tidyr::fish_encounters, "chr", return_tbl = FALSE)
require(dplyr)
require(tidyr)
get_column_class(dplyr::starwars, class = "chr")
get_column_class(dplyr::starwars, class = "chr", return_tbl = FALSE)
get_column_class(dplyr::starwars, class = "list")
get_column_class(dplyr::starwars, class = "list", return_tbl = FALSE)
get_column_class(tidyr::fish_encounters, "chr")
get_column_class(tidyr::fish_encounters, "chr", return_tbl = FALSE)
}
