% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{get_facet_checks_vec}
\alias{get_facet_checks_vec}
\title{Facet checks (helper)}
\usage{
get_facet_checks_vec(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{type of column to return.}
}
\value{
TRUE/FALSE if facet variable (< 6 levels)
}
\description{
Facet checks (helper)
}
\examples{
require(NHANES)
get_facet_checks_vec(NHANES::NHANES$Education, type = "fct")
levels(NHANES::NHANES$Education)
get_facet_checks_vec(NHANES::NHANES$MaritalStatus, type = "fct")
levels(NHANES::NHANES$MaritalStatus)
}
