% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ui_inputs}
\alias{get_ui_inputs}
\title{Get UI inputs}
\usage{
get_ui_inputs(app_data)
}
\arguments{
\item{app_data}{dataset for app (\code{data.frame} or \code{tibble})}
}
\value{
zeallot assignment (\verb{\%<-\%}) with input character vector on LHS and
list of names by type on the RHS
}
\description{
Get UI inputs
}
\examples{
require(dplyr)
get_ui_inputs(dplyr::starwars)
}
