% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_var_types}
\alias{get_var_types}
\title{Get data variable types (as list)}
\usage{
get_var_types(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
list of vectors by type (\code{is_logical}, \code{is_double}, \code{is_integer},
\code{is_character}, \code{is_factor}, \code{is_facet_var}, \code{is_date}, \code{is_POSIXct},
\code{is_POSIXlt})
}
\description{
Get data variable types (as list)
}
\examples{
require(dplyr)
get_var_types(dplyr::starwars)
get_var_types(dplyr::storms)
# great with zealott!
require(zeallot)
c(dbl_vars, int_vars,
  fct_vars, facet_vars) \%<-\% get_var_types(penguins)
dbl_vars
int_vars
fct_vars
facet_vars
}
