% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_base.R
\name{gg_base}
\alias{gg_base}
\title{Base plot (ggplot2)}
\usage{
gg_base(df, x_var, y_var)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable}

\item{y_var}{y variable}
}
\value{
plot object
}
\description{
Base plot (ggplot2)
}
\examples{
require(testPkgApp)
movies <- testPkgApp::movies
gg_base(df = movies, x_var = "critics_score", y_var = "imdb_rating")
}
