% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_points.R
\name{gg_points}
\alias{gg_points}
\title{Point plot (ggplot2)}
\usage{
gg_points(df, x_var, y_var, col_var, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{col_var}{color variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{...}{other arguments passed to \code{ggplot2::geom_point()}, outside of \code{ggplot2::aes()}}
}
\value{
A \code{ggplot2} plot object
}
\description{
Point plot (ggplot2)
}
\examples{
require(testPkgApp)
movies <- testPkgApp::movies
gg_points(
  df = movies,
  x_var = "critics_score",
  y_var = "imdb_rating",
  col_var = "critics_rating",
  alpha = 1 / 3, size = 2
)
}
