% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_points_facet.R
\name{gg_points_facet}
\alias{gg_points_facet}
\title{Point plot with facets (ggplot2)}
\usage{
gg_points_facet(df, x_var, y_var, col_var, facet_var, ...)
}
\arguments{
\item{df}{input dataset (tibble or data.frame)}

\item{x_var}{x variable (supplied to \code{ggplot2::aes(x = )})}

\item{y_var}{y variable (supplied to \code{ggplot2::aes(y = )})}

\item{col_var}{color variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{facet_var}{facet variable (supplied to \code{ggplot2::geom_point(ggplot2::aes(color = ))})}

\item{...}{other arguments passed to (\code{ggplot2::facet_wrap(vars())})}
}
\value{
A \code{ggplot2} plot object
}
\description{
Point plot with facets (ggplot2)
}
\examples{
require(palmerpenguins)
gg_points_facet(
  df = palmerpenguins::penguins,
  x_var = "bill_length_mm",
  y_var = "flipper_length_mm",
  col_var = "island",
  facet_var = "species",
  alpha = 1 / 3,
  size = 2
)
gg_points_facet(
  df = palmerpenguins::penguins,
  x_var = "bill_length_mm",
  y_var = "flipper_length_mm",
  col_var = "island",
  facet_var = NULL,
  alpha = 1 / 3,
  size = 2
)
gg_points_facet(
  df = palmerpenguins::penguins,
  x_var = "bill_length_mm",
  y_var = "flipper_length_mm",
  col_var = NULL,
  facet_var = NULL,
  alpha = 1 / 3,
  size = 2
)
}
