% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_binary_checks_df}
\alias{make_binary_checks_df}
\title{Create vector of binary columns from data.frame or tibble (helper)}
\usage{
make_binary_checks_df(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of binary columns
}
\description{
Create vector of binary columns from data.frame or tibble (helper)
}
\examples{
require(palmerpenguins)
require(dplyr)
bins <- make_binary_checks_df(
            df = dplyr::select(palmerpenguins::penguins,
                          dplyr::where(is.factor)),
            type = "fct")
bins
}
