% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_binary_cols_vec}
\alias{make_binary_cols_vec}
\title{Get binary columns from data}
\usage{
make_binary_cols_vec(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
vector of binary column names
}
\description{
Get binary columns from data
}
\examples{
require(dplyr)
make_binary_cols_vec(dplyr::starwars)
}
