% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_col_types_vec}
\alias{make_col_types_vec}
\title{Data column type}
\usage{
make_col_types_vec(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return.
One of:
\itemize{
\item "log": logical
\item "dbl": double
\item "int": integer
\item "chr": character
\item "fct": factor
\item "lst": list
\item "date": date (from \code{lubridate::is.Date})
\item "posixct": date (from \code{lubridate::is.POSIXct})
\item "posixlt": date (from \code{lubridate::is.POSIXlt})
\item "posix": date (from \code{lubridate::is.POSIXt})
\item "facet": facet variables (less than six levels)
\item "binary": binary variables (two levels)
}}
}
\value{
vector of names from df matching \code{type}
}
\description{
Data column type
}
\examples{
require(palmerpenguins)
require(dplyr)
make_col_types_vec(df = palmerpenguins::penguins, type = 'dbl')
make_col_types_vec(df = palmerpenguins::penguins, type = 'int')
}
