% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_facet_checks_df}
\alias{make_facet_checks_df}
\title{Create vector of facet columns from data (helper)}
\usage{
make_facet_checks_df(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of facet columns (< 6 levels)
}
\description{
Create vector of facet columns from data (helper)
}
\examples{
require(NHANES)
make_facet_checks_df(NHANES::NHANES, type = "fct")
}
