% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_title.R
\name{make_plot_title}
\alias{make_plot_title}
\title{Make plot title}
\usage{
make_plot_title(x, y, color)
}
\arguments{
\item{x}{x variable}

\item{y}{y variable}

\item{color}{color variable}
}
\value{
String for plot title
}
\description{
Make plot title
}
\examples{
make_plot_title(
  x = "imdb_rating",
  y = "audience_score",
  color = "mpaa_rating"
)
}
