% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_ui_inputs}
\alias{make_ui_inputs}
\title{Make UI inputs}
\usage{
make_ui_inputs(app_data)
}
\arguments{
\item{app_data}{dataset for app (\code{data.frame} or \code{tibble})}
}
\value{
zeallot assignment (\verb{\%<-\%}) with input character vector on LHS and
list of names by type on the RHS
}
\description{
This is meant to be used in the console--it generates the code for assigning
the elements from a list into a collection of vectors.
}
\examples{
require(dplyr)
make_ui_inputs(dplyr::starwars)
}
