% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_var_input.R
\name{mod_var_input_ui}
\alias{mod_var_input_ui}
\title{Import variables UI module}
\usage{
mod_var_input_ui(id, app_data)
}
\arguments{
\item{id}{module id}

\item{app_data}{data for application}
}
\value{
shiny UI module
}
\description{
Import variables UI module
}
