% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{non_fct_binary_vars}
\alias{non_fct_binary_vars}
\title{Non-factor binary variables}
\usage{
non_fct_binary_vars(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of character or logical column names with two unique values
}
\description{
Non-factor binary variables
}
