% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{num_app_inputs}
\alias{num_app_inputs}
\title{Numeric app inputs}
\usage{
num_app_inputs(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
integer and double column names
}
\description{
Numeric app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
num_app_inputs(palmerpenguins::penguins)
num_app_inputs(dplyr::starwars)
}
