% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_app_inputs.R
\name{test_app_inputs}
\alias{test_app_inputs}
\title{Demo \verb{*app_inputs()} functions}
\usage{
test_app_inputs(data)
}
\arguments{
\item{data}{\code{app_data} can be any dataset (automatically loads with \code{NHANES::NHANES})}
}
\value{
demo shiny application
}
\description{
Demo \verb{*app_inputs()} functions
}
\examples{
require(testPkgApp)
# test_app_inputs()
}
