% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPkgApp-package.R
\docType{package}
\name{unitTestAppPkg-package}
\alias{unitTestAppPkg}
\alias{unitTestAppPkg-package}
\title{unitTestAppPkg: Unit Test Your Shiny App Package}
\description{
This shiny app was developed as a package using devtools and usethis.
}
\keyword{internal}
