% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{vec_binary_checks}
\alias{vec_binary_checks}
\title{Binary checks (vectors)}
\usage{
vec_binary_checks(x, type)
}
\arguments{
\item{x}{a vector}

\item{type}{type of column to return.}
}
\value{
TRUE/FALSE if binary
}
\description{
Binary checks (vectors)
}
\examples{
require(palmerpenguins)
vec_binary_checks(palmerpenguins::penguins$sex, type = "fct")
}
