% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{binary_app_inputs}
\alias{binary_app_inputs}
\title{Binary app inputs}
\usage{
binary_app_inputs(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
named vector of binary column names
}
\description{
Binary app inputs
}
\examples{
require(palmerpenguins)
require(dplyr)
binary_app_inputs(palmerpenguins::penguins)
binary_app_inputs(dplyr::starwars)
}
