% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{col_type_list}
\alias{col_type_list}
\title{Get data variable types (as list)}
\usage{
col_type_list(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
list of vectors by type:
\itemize{
\item \code{logical_vars}
\item \code{double_vars}
\item \code{integer_vars}
\item \code{character_vars}
\item \code{factor_vars}
\item \code{date_vars}
\item \code{posixct_vars}
\item \code{posixlt_vars}
\item \code{posixt_vars}
\item \code{list_vars}
}
}
\description{
Get data variable types (as list)
}
\examples{
require(palmerpenguins)
col_type_list(palmerpenguins::penguins)
}
