% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_inputs.R
\name{facet_app_inputs}
\alias{facet_app_inputs}
\title{Facet app inputs}
\usage{
facet_app_inputs(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of factor or character column names with < 6 unique levels
}
\description{
Facet app inputs
}
\section{Variables to use for facets}{


This function is designed to quickly determine which variables have an
appropriate number of categorical levels for using \code{ggplot2::facet_wrap()}
or \code{ggplot2::facet_grid()}
}

\examples{
require(dplyr)
require(NHANES)
facet_app_inputs(df = dplyr::starwars)
facet_app_inputs(df = NHANES::NHANES)

str(dplyr::select(
  dplyr::starwars,
  dplyr::all_of(facet_app_inputs(df = dplyr::starwars))
))
str(dplyr::select(
  NHANES::NHANES,
  dplyr::all_of(facet_app_inputs(df = NHANES::NHANES))
))
}
