% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{facet_vars}
\alias{facet_vars}
\title{Facet variables (as vector)}
\usage{
facet_vars(df)
}
\arguments{
\item{df}{a data frame or tibble}
}
\value{
vector of factor or character columns with less than six levels
}
\description{
Facet variables (as vector)
}
\examples{
require(tibble)
df_test <- tibble::tibble(
  a = sample(c(TRUE, FALSE), 10, TRUE),
  b = c(1.5:10.5),
  c = c(1:10),
  d1 = rep(LETTERS[1:5], times = 2),
  d2 = LETTERS[1:10],
  e = sample(c(
    Sys.Date(), Sys.Date() + 1,
    Sys.Date() + 2, Sys.Date() + 3,
    Sys.Date() + 4
  ), size = 10, replace = TRUE),
  f1 = factor(rep(letters[1:5], times = 2),
    levels = letters[1:5]
  ),
  f2 = factor(letters[1:10],
    levels = letters[1:10]
  )
)
facet_vars(df_test)
}
