% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{get_binary_cols}
\alias{get_binary_cols}
\title{Get binary columns from data}
\usage{
get_binary_cols(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
vector of binary column names
}
\description{
Get binary columns from data
}
\examples{
require(dplyr)
get_binary_cols(dplyr::starwars)
}
