% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_col_type_df}
\alias{get_col_type_df}
\title{Get column types (helper)}
\usage{
get_col_type_df(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return.}
}
\value{
columns of \code{type}
}
\description{
Get column types (helper)
}
\examples{
require(dplyr)
get_col_type_df(dplyr::starwars, type = "chr")
get_col_type_df(dplyr::starwars, type = "list")
}
