% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_binary_checks_df}
\alias{make_binary_checks_df}
\title{Check for binary columns (helper)}
\usage{
make_binary_checks_df(df, type)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}

\item{type}{type of column to return}
}
\value{
vector of binary columns
}
\description{
Check for binary columns (helper)
}
\examples{
require(palmerpenguins)
make_binary_checks_df(palmerpenguins::penguins, type = "fct")
}
