% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_app_inputs.R
\name{make_facet_cols_vec}
\alias{make_facet_cols_vec}
\title{Get facet columns from data}
\usage{
make_facet_cols_vec(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}}
}
\value{
a vector of factor or character column names with < 6 unique levels
}
\description{
Get facet columns from data
}
\section{Variables to use for facets}{


This function is designed to quickly determine which variables have an
appropriate number of categorical levels for using \code{ggplot2::facet_wrap()}
or \code{ggplot2::facet_grid()}
}

\examples{
require(dplyr)
require(NHANES)
make_facet_cols_vec(df = dplyr::starwars)
make_facet_cols_vec(df = NHANES::NHANES)

str(dplyr::select(dplyr::starwars,
  dplyr::all_of(make_facet_cols_vec(df = dplyr::starwars))))
str(dplyr::select(NHANES::NHANES,
  dplyr::all_of(make_facet_cols_vec(df = NHANES::NHANES))))
}
