% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testPkgApp-package.R
\docType{package}
\name{testPkgApp-package}
\alias{testPkgApp}
\alias{testPkgApp-package}
\title{testPkgApp: Build Your Shiny App Like a Package}
\description{
This shiny app was developed as a package using devtools and usethis.
}
\keyword{internal}
